+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.3.0
+- Splitted REMREM Publish component into cli, service and common.
+
+## 0.2.9
+- Updated the RemRem shared version to latest.
+
 ## 0.2.8
 - Implemented RabbitMQ connection retry in Service.